+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Add `fxtest.App`, which redirects log output to the user's `testing.TB` and
   provides some lifecycle helpers.
 - **[Breaking]** Alter lifecycle hooks to accept a context.
+- Add an `Err` method to retrieve the underlying errors during the dependency
+  graph construction. The same error is also returned from `Start`.
+- Graph resolution now happens as part of `fx.New`, rather than at the beginning
+  of `app.Start`. This allows inspection of the graph errors through `app.Err()`
+  before the decision to start the app.
 
 ## v1.0.0-rc1 (20 Jun 2017)